Merged PR 10: Venue landing page UI fit and finish
The tabs in venue landing must be centered properly on mobile. The About tab should be left most , Reviews tab should be in center and Add Review tab should be in the right. All tabs must be displayed without having to scroll.

The 'More Info' button should appear in the About Tab.

 not able to close the "More hours" dialog after it displays.